requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Werkzeug>=3.0.1
python-dotenv>=1.0.0
pytz>=2024.1
reportlab>=4.1.0
openpyxl>=3.1.2
Pillow>=10.3.0
